[ pull request ][ GW2App : develop ]  

1. 整理的還不錯，給你鼓勵鼓勵！
2. 有個習慣要改，就是 .m 實作曾的 function 分類還是太混亂

+++ GW2App/GW2App/Tools/Constants.h
@@ -10,7 +10,7 @@
 #define GW2BroH_Constants_h
 
 #pragma mark - Enum
-typedef enum : NSInteger{
+typedef enum : int{
     EnumTabBarIndexStartMenu = 0,
     EnumTabBarIndexWorldBoss = 10,
     EnumTabBarIndexItems = 11,
